fix(db): port the three SQLite-only bounty functions, and catch the class
`bounties` returned "decoding column 4: mismatched types; Rust type `i64` (as SQL
type `INT8`) is not compatible with SQL type `INT4`" on PostgreSQL. `amount` is
INTEGER there, so both bounty reads now widen in SQL -- `amount::bigint` -- which
lets one row type serve both backends. `list_bounties` and `create_bounty` had no
PostgreSQL arm at all; `claim_bounty` and `seed::run` had the same problem, the
latter reached from `lorehaven seed` on the CLI, so a seeded PostgreSQL
deployment panicked on the bare string "sqlite in seed".

`check-pg-arm-uses-sqlite-pool.py` could not see any of this: it flags a
`sqlite_pool()` sitting inside a `Backend::Postgres` arm, and these functions have
no such arm. It now also reports functions that use the pool while naming no
backend at all, which is the shape that passes every SQLite test and misbehaves
under LOREHAVEN_TEST_PG_URL with an error about a column type. Two of the four it
found are correct as written -- helpers called from one arm -- and are exempted
by the `_sqlite` naming convention this repo already uses for them.

+++ crates/db/src/bounties.rs
@@ -21,6 +21,9 @@ use serde_json::json;
 ///
 /// Named because the row is read in several places and the ten-element tuple is
 /// otherwise unreadable. The element order must match the SELECT lists exactly.
+/// The two integer columns are `INTEGER` in PostgreSQL and widen to `i64` in
+/// the query. They decode as `i64` here only because the PostgreSQL arm casts;
+/// SQLite hands back `i64` for the same columns, so one type serves both.
 type BountyRow = (
     String,
     String,
@@ -223,7 +226,8 @@ pub async fn fetch_bounty(db: &Database, id: &str) -> Result<Option<Bounty>, sql
         crate::Backend::Postgres => {
             let pool = db.postgres_pool().expect("postgres handle");
             sqlx::query_as(
-                "SELECT id, type, job_kind, terms, amount, funded_amount, state, created_by, created_at, activated_at
+                "SELECT id, type, job_kind, terms, amount::bigint, funded_amount::bigint,
+                        state, created_by, created_at, activated_at
                  FROM bounties WHERE id = $1",
             )
             .bind(id)
